feat: Add login functionality

+++ vagrant/my_skills.py
@@ -7,7 +7,6 @@ APP.secret_key = "VerySecretKey"
 # imports to connect script to database
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-
 from database_setup import SkillTable, Base, CourseTable
 
 # connect script to my_skills.db database
@@ -16,11 +15,26 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
+# imports to add login functionality
+from flask import session as login_session
+import random
+import string
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
+
+
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "my_skills App"
+
 
 # set up the routing for each page
 
 @APP.route('/')
-@APP.route('/home/')
 def home_page():
     """routing for home page"""
     skill_lst = session.query(SkillTable).all()
@@ -40,6 +54,8 @@ def edit_skill(skill):
     """routing to edit a specific skill"""
     skill_item = session.query(SkillTable).filter_by(name=skill).one()
     if request.method == 'POST':
+        if 'username' not in login_session:
+            return redirect('/login/')
         if request.form['name']:
             skill_item.name = request.form['name']
         session.add(skill_item)
@@ -54,6 +70,8 @@ def delete_skill(skill):
     skill_item = session.query(SkillTable).filter_by(name=skill).one()
     course_lst = session.query(CourseTable).filter_by(skill_id=skill_item.id).all()
     if request.method == 'POST':
+        if 'username' not in login_session:
+            return redirect('/login/')
         for x in course_lst:
             session.delete(x)
         session.delete(skill_item)
@@ -65,19 +83,14 @@ def delete_skill(skill):
 @APP.route('/skill/new/', methods=['GET', 'POST'])
 def new_skill():
     """routing to create a new category"""
+    if 'username' not in login_session:
+        return redirect('/login/')
     if request.method == 'POST':
         add_skill = SkillTable(name=request.form['name'])
         session.add(add_skill)
         session.commit()
         return redirect(url_for('home_page'))
     return render_template('new_skill.html')
-
-
-@APP.route('/login/')
-def login_page():
-    """routing for the designated login page"""
-    return render_template('login.html')
-
 
 @APP.route('/<skill>/<int:course>')
 def course_page(skill, course):
@@ -91,6 +104,8 @@ def course_page(skill, course):
 def new_course(skill):
     """routing for a new course in skillset"""
     skill_item = session.query(SkillTable).filter_by(name=skill).one()
+    if 'username' not in login_session:
+        return redirect('/login/')
     if request.method == 'POST':
         iterator = 1
         switch = True
@@ -114,6 +129,8 @@ def edit_course(skill, course):
     """routing to edit a course in a skillset"""
     skill_item = session.query(SkillTable).filter_by(name=skill).one()
     course_item = session.query(CourseTable).filter_by(id=course).one()
+    if 'username' not in login_session:
+        return redirect('/login/')
     if request.method == 'POST':
         if request.form['name']:
             course_item.name = request.form['name']
@@ -134,6 +151,8 @@ def delete_course(skill, course):
     """routing to delete an item in a category"""
     skill_item = session.query(SkillTable).filter_by(name=skill).one()
     course_item = session.query(CourseTable).filter_by(id=course).one()
+    if 'username' not in login_session:
+        return redirect('/login/')
     if request.method == 'POST':
         session.delete(course_item)
         session.commit()
@@ -155,6 +174,100 @@ def course_api(skill):
 
 
 
+
+
+@APP.route('/login/')
+def show_login():
+    """routing for the designated login page"""
+    # Create anti-forgery state token
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)\
+        for x in range(32))
+    login_session['state'] = state
+    return render_template('login.html', STATE=state)
+
+
+@APP.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1].decode())
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print("Token's client ID does not match app's.")
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # check if user is already connected
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print("done!")
+    return output
